do not delete new trained model, and drop several tests that load trained model
